Fix scatter_add_ int64 index requirement and auto-tune error handling
- Cast scatter_add_ indices to int64 in cutlass_grouped_hybrid and
  explicit_gemm_grouped (torch 2.7+ requires int64 for scatter indices)
- Add torch.cuda.synchronize() after each auto-tune candidate to catch
  async CUDA errors before they corrupt subsequent operations

+++ warpconvnet/nn/functional/sparse_conv/detail/explicit.py
@@ -167,7 +167,7 @@ def _explicit_gemm_forward_grouped(
         result_flat = result.view(B * max_m, C_out)
         output_feature_tensor.scatter_add_(
             0,
-            cat_out.unsqueeze(1).expand(-1, C_out),
+            cat_out.unsqueeze(1).expand(-1, C_out).long(),
             result_flat[flat_idx].to(dtype=output_feature_tensor.dtype),
         )
 
@@ -250,7 +250,7 @@ def _explicit_gemm_backward_grouped(
         grad_in_result_flat = grad_in_result.view(B * max_m, C_in)
         grad_in_features.scatter_add_(
             0,
-            cat_in.unsqueeze(1).expand(-1, C_in),
+            cat_in.unsqueeze(1).expand(-1, C_in).long(),
             grad_in_result_flat[flat_idx].to(dtype=grad_in_features.dtype),
         )
